Move new added file to hysteresis folder
